+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 - Input gain and output gain controls for the equalizer.
 - Removed a small memory leak(nothing critical).
 
+### Fixed
+- Compressor and gate plugins knee parameter was always at its default value.
+ Fixed that.
+
 ### General notes
 - The limiter plugin in latest Calf Studio release (0.90) has a memory leak.
 I reported it to Calf developers and it was fixed. See